Enable Filter on Servlets
Previously, this implementation was only contributed to Spring Boot
applications, although it would be for both JARs and WARs.  This left straight
WAR applications (non Spring Boot) without the functionality that this project
provides.  This change adds a Servlet 3.0 ServletContextInitializer that adds
the filter to all requests in a Servlet container enabling the mapping
behavior for WAR applications.

[resolves #531]